Move downloading code to wn._download
